+++ packages/jest-jasmine2/src/reporter.ts
@@ -113,15 +113,19 @@ export default class Jasmine2Reporter implements Reporter {
     return this._resultsPromise;
   }
 
-  private _addMissingMessageToStack(stack: string, message?: string) {
+  private _addMissingMessageToStack(stack: unknown, message?: string) {
     // Some errors (e.g. Angular injection error) don't prepend error.message
     // to stack, instead the first line of the stack is just plain 'Error'
     const ERROR_REGEX = /^Error:?\s*\n/;
 
-    if (stack && message && !stack.includes(message)) {
-      return message + stack.replace(ERROR_REGEX, '\n');
+    if (typeof stack === 'string') {
+      if (stack && message && !stack.includes(message)) {
+        return message + stack.replace(ERROR_REGEX, '\n');
+      }
+      return stack;
+    } else {
+      return `${stack}`;
     }
-    return stack;
   }
 
   private _extractSpecResults(